requirements.txt
openenv-core[core]>=0.2.0
trl>=0.12.0
transformers>=4.36.0
datasets>=2.14.0
vllm>=0.6.0
torch>=2.0.0
pydantic>=2.0.0
numpy>=1.24.0
